requirements.txt
pillow
bleak
python-dotenv
requests